pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "digit-classifier-app"
dynamic = ["version"]
description = "Handwritten Digit Classifier App"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = "MIT"
dependencies = [
    "streamlit >=0.79.0",
    "torch >=1.9.0",
    "numpy >=1.20.0",
    "matplotlib >=3.4.0",
    "scikit-learn >=0.24.0",
    "Pillow >=8.2.0"
]
requires-python = ">=3.10,<3.11"

[tool.hatch.version]
path = "python/digit_classifier_app/__init__.py"

[tool.hatch.build.targets.wheel]
packages = [
    "python/digit_classifier_app",
]

[project.scripts]
digit-classifier = "digit_classifier_app.main:main"

[tool.isort]
src_paths = ["python/"]
profile = "black"
